Changed the structure of the project. Button, Info, Menu, Settings, SubMenu and worldCreator classes were moved to the GUI folder. Character, Game and Unit classes were moved to the core folder
