+++ src/Projectile/Rock.cpp
@@ -1,5 +1,5 @@
 #include "Rock.h"
-#include "../Character.h"
+#include "../core/Character.h"
 Rock::Rock(sf::Vector2f p, sf::Vector2f dest, float dmg, bool attackMons) : Projectile("Rock", dmg, sf::seconds(2),PHYSIC,p,{80.f,80.f},dest,350.f,attackMons){
 createBodyPart({80.f,80.f},{0.f,0.f},sf::Color{40,40,40});
 disappearOnImpact=true;
