+++ src/core/Character.cpp
