+++ src/core/Unit.h
@@ -2,7 +2,7 @@
 #define UNIT_H
 
 #include <SFML/Graphics.hpp>
-#include "Utilities/Utils.h"
+#include "../Utilities/Utils.h"
 #include <cmath>
 #include <memory>
 #include <iostream>
