+++ src/monsters/GiantSpider.h
@@ -1,7 +1,7 @@
 #ifndef GIANTSPIDER_H
 #define GIANTSPIDER_H
 
-#include "../Monster.h"
+#include "Monster.h"
 
 class GiantSpider : public Monster{
 public:
