+++ src/monsters/Troll.h
@@ -1,7 +1,7 @@
 #ifndef TROLL_H
 #define TROLL_H
 
-#include "../Monster.h"
+#include "Monster.h"
 
 class Troll : public Monster{
 public:
