+++ src/monsters/Zombie.h
@@ -1,7 +1,7 @@
 #ifndef ZOMBIE_H
 #define ZOMBIE_H
 
-#include "../Monster.h"
+#include "Monster.h"
 
 class Zombie : public Monster{
     public:
